Рефакторинг компонента для отрисовки и загрузки аватара пользователя
