+++ client/src/Styles/Profile.less
@@ -17,6 +17,7 @@
             margin-bottom: 30px;
             height: 150px;
             width: 150px;
+            min-width: 150px;
             border: 2px solid #dfdfe4;
             border-radius: 100px;
